int -> uint16_t

+++ main.cpp
@@ -5,7 +5,7 @@
 /********************************************/
 
 // prime ceiling
-constexpr int PS_CEIL = 1000;
+constexpr uint16_t PS_CEIL = 1000;
 
 // pipe output
 // CURRENTLY DOES NOT WORK, KEEP TO FALSE
@@ -13,15 +13,15 @@ constexpr bool PS_PIPED = false;
 
 /********************************************/
 
-std::vector<int> getPrimes(const int &ceiling)
+std::vector<uint16_t> getPrimes(const uint16_t& ceiling)
 {
-	std::vector<std::pair<int, bool>> primes;
-	std::vector<int> primesReturn;
+	std::vector<std::pair<uint16_t, bool>> primes;
+	std::vector<uint16_t> primesReturn;
 	// any number past this doesnt need to be processed in prime loop
-	int maxToAdd = (int) sqrt(ceiling);
+	uint16_t maxToAdd = (uint16_t) sqrt(ceiling);
 	
 	primes.push_back({1,false});
-	for (int i = 2; i <= ceiling; i++) primes.push_back({i,true});
+	for (uint16_t i = 2; i <= ceiling; i++) primes.push_back({i,true});
 
 	for (auto i : primes) {
 		// if number is prime
@@ -29,7 +29,7 @@ std::vector<int> getPrimes(const int &ceiling)
 			primesReturn.push_back(i.first);
 
 			if (i.first < maxToAdd) {
-				for (int x = (i.first * 2); x <= ceiling; x += i.first) {
+				for (uint16_t x = (i.first * 2); x <= ceiling; x += i.first) {
 					if (x > ceiling) break;
 					primes[x-1].second = false;
 				}
@@ -48,7 +48,7 @@ int main()
 		std::cout.rdbuf(out.rdbuf());
 	}
 	auto primes = getPrimes(PS_CEIL);
-	for (int i : primes) std::cout << i << " ";
+	for (uint16_t i : primes) std::cout << i << " ";
 
 	return 0;
 }
